fix(section-divider): drop None option, add aria-hidden, fix range reset, dedupe style builder
