+++ src/blocks/section-divider/edit.js
@@ -3,9 +3,9 @@
  *
  * Renders the identical classed inner div as save.js so the shape paints
  * live on the canvas: same outer/inner structure, same emit-only-when-
- * differs inline style contract. Inspector controls live in the Theme-3
- * Settings panel (shape/height/width/flip) plus the shared color group
- * (fill).
+ * differs inline style contract (both derived from ./utils so they can't
+ * drift). Inspector controls live in the Theme-3 Settings panel
+ * (shape/height/width/flip) plus the shared color group (fill).
  *
  * @since 2.7.0
  */
@@ -18,7 +18,7 @@ import {
 	__experimentalColorGradientSettingsDropdown as ColorGradientSettingsDropdown,
 } from '@wordpress/block-editor';
 import { useBlockColors } from '../../hooks';
-import { convertColorToCSSVar } from '../../utils/convert-preset-to-css-var';
+import { getDividerStyle, getDividerShapeClass } from './utils';
 import SectionDividerControls from './components/SectionDividerControls';
 
 export default function SectionDividerEdit( {
@@ -26,41 +26,16 @@ export default function SectionDividerEdit( {
 	setAttributes,
 	clientId,
 } ) {
-	const { shape, height, width, flipX, flipY, fillColor } = attributes;
-
 	const { settings: colorSettings, colorGradientSettings } = useBlockColors( {
 		attributes,
 		setAttributes,
 		entries: [ { label: __( 'Fill', 'designsetgo' ), attribute: 'fillColor' } ],
 	} );
 
-	// Mirrors save.js exactly — same vars, same emit-only-when-differs guards —
-	// so the editor canvas matches the frontend output byte for byte.
-	const style = {};
-
-	if ( fillColor ) {
-		style[ '--dsgo-section-divider-fill' ] =
-			convertColorToCSSVar( fillColor );
-	}
-
-	if ( typeof height === 'number' ) {
-		style[ '--dsgo-shape-height' ] = `${ height }px`;
-	}
-
-	if ( width !== 100 ) {
-		style[ '--dsgo-shape-width' ] = `${ width }%`;
-	}
-
-	if ( flipX ) {
-		style[ '--dsgo-shape-flip-x' ] = -1;
-	}
-
-	if ( flipY ) {
-		style[ '--dsgo-shape-flip-y' ] = -1;
-	}
-
-	const shapeClass =
-		shape === 'inherit' ? 'is-shape-inherit' : `is-shape-${ shape }`;
+	// Derived from the same ./utils helpers save.js uses, so the editor
+	// canvas matches the frontend output byte for byte.
+	const style = getDividerStyle( attributes );
+	const shapeClass = getDividerShapeClass( attributes.shape );
 
 	const blockProps = useBlockProps();
 
@@ -87,6 +62,7 @@ export default function SectionDividerEdit( {
 				<div
 					className={ `dsgo-section-divider__shape dsgo-shape-divider ${ shapeClass }` }
 					style={ Object.keys( style ).length ? style : undefined }
+					aria-hidden="true"
 				/>
 			</div>
 		</>
